Assign Issue Updated

+++ yashissuetrackersystem/yashissuetrackersystem-core/src/main/java/com/yash/yits/dao/IssueDao.java
@@ -59,7 +59,7 @@ public interface IssueDao {
 
 	public List<Member> getMemberList();
 
-	public void assignIssue(Issue issue, int fetchId);
+	public void assignIssue(Issue issue,Long fetchId);
 	public Issue fetchIssueDetailsConv(int id);
 
 	public List<Member> getMemberListConv();
